# src/config/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perfume Shop API",
    "description": "RESTful API for a perfume and cologne e-commerce platform with multi-variant product support, allowing flexible volume-based pricing and weight-based sales",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourdomain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/item": {
      "post": {
        "tags": [
          "Cart"
        ],
        "summary": "Add item to cart",
        "description": "Add a product variant (volume) to the shopping cart",
        "responses": {
          "200": {
            "description": "Item added to cart successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Item added to cart successfully."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalItemToCart": {
                          "type": "integer",
                          "example": 3
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or stock limit exceeded"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Product or volume not found or inactive"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011",
                    "description": "Product ID (MongoDB ObjectId)"
                  },
                  "volumeId": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439012",
                    "description": "Volume/variant ID (MongoDB ObjectId)"
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 1,
                    "description": "Quantity (1-10)"
                  }
                },
                "required": [
                  "productId",
                  "volumeId",
                  "quantity"
                ]
              }
            }
          }
        }
      }
    },
    "/item/{itemId}/increase": {
      "patch": {
        "tags": [
          "Cart"
        ],
        "summary": "Increase item quantity",
        "description": "Increase the quantity of an item in the cart",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart item ID (MongoDB ObjectId)"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart item quantity increased successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "itemId": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Quantity update exceeds allowed limits"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Item not found in cart"
          },
          "409": {
            "description": "Invalid item ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "example": 1,
                    "description": "Quantity to increase by (1-10)"
                  }
                },
                "required": [
                  "quantity"
                ]
              }
            }
          }
        }
      }
    },
    "/item/{itemId}/decrease": {
      "patch": {
        "tags": [
          "Cart"
        ],
        "summary": "Decrease item quantity",
        "description": "Decrease the quantity of an item in the cart (removes if quantity becomes 0)",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart item quantity decreased successfully"
          },
          "400": {
            "description": "Quantity update exceeds allowed limits"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Item not found in cart"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "example": 1,
                    "description": "Quantity to decrease by"
                  }
                },
                "required": [
                  "quantity"
                ]
              }
            }
          }
        }
      }
    },
    "/item/{itemId}": {
      "delete": {
        "tags": [
          "Cart"
        ],
        "summary": "Remove item from cart",
        "description": "Remove a specific item from the shopping cart",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart item removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "itemId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Item not found in cart"
          },
          "409": {
            "description": "Invalid item ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/{slug}/reactivate": {
      "patch": {
        "tags": [
          "Admin - Categories"
        ],
        "summary": "Reactivate a category",
        "description": "Activate a deactivated category",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Category slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Category reactivated successfully"
          },
          "400": {
            "description": "Category already active"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/{userId}/reactivate": {
      "patch": {
        "tags": [
          "Admin - Users"
        ],
        "summary": "Reactivate a user",
        "description": "Reactivate a deactivated user account (Admin only)",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User reactivated successfully"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Invalid user ID format"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    }
  }
}